Use a single evaluation enviroment for rules. (#84)
The rule extractor interprets Skylark code as Python, in an
environment with most Skylark globals and primitives stubbed out. It
was doing so in two environments, one for globals and one for locals.
But the Python documentation has this to say:

> If two separate objects are passed as globals and locals, the code
> will be executed as if it were embedded in a class definition.

This means that using two environments can have suprising results. One
of which is that calling recursive functions don't work, whereas it
works in Skylark or at the Python top-level. This also means that
calling a function `foo()` that itself depends on a function `bar()`
in the same Skylark file doesn't work.

So we instead `exec ...` in a single environment. We infer new
definitions contained in the input file by diffing the hold state and
the new state of the environment.