Fix runfile lookup in skydoc/main.py (#108)
* Fix runfile lookup in skydoc/main.py

`skydoc/main.py` now uses Bazel's built-in
runfiles library to look up data-dependencies.

This commit introduces one hack: we use
os.path.dirname (actually: posixpath.dirname) on a
path returned by the runfiles library.

This is wrong, because the runfiles path does not
necessarily correlate with the actual path of the
file, i.e. the path that the runfiles-lookup
library returns. For example, if //foo:a and
//foo:generated-b are both data-dependencies,
they'll both be under the "foo"
runfiles-directory, but will be in different real
paths. This is fine on Linux because Bazel
generates a runfiles tree with symlinks, but
breaks on Windows because there Bazel only
generates a runfiles manifest.

Fortunately, for now, all files under
`skydoc/templates` are source files, so this hack
works. If anyone ever adds generated files to
//skydoc/templates:templates, this hack will
break.

See https://github.com/bazelbuild/skydoc/issues/58