Fix the generated links to things with CamelCase names
The URL fragment `something.html#xyz` links to an element with `id=xyz`,
not `id="#xyz"` like was being generated. In most cases, this didn't
matter, because `some_thing` generates a heading with `id=some_thing` in
addition to the `<a>` which didn't work. However, `SomeThing` produces a
heading with `id=something`, which doesn't match, so clicking the
relevant links does not scroll the page to the intended location.

If anybody was linking to `something.html##xyz`, those won't go to the
intended location any more, but that doesn't seem like intended behavior
to me.