Update bazel and rules_pkg
If you have a bzl file that contains
`@bazel_tools//tools/cpp/toolchain_utils.bzl`, after this commit
https://github.com/bazelbuild/bazel/commit/95f0135fbac934be696676f37e0285b2d311603a
stardoc stopped working because it didn't know about this API, leading
to this crash https://buildkite.com/bazel/rules-apple-darwin/builds/5829#0181f96c-ff6c-449e-936e-c32530bd3360

I assume this is because some of the stardoc internals pulls in the
older version of `ConfigStarlarkCommonApi.java`. This bumps the shas in
the first commit, and rebuilds them in the second commit. Feel free to
drop my rebuilds and do a more trusted rebuild.