List maven artifacts in MODULE.bazel to avoid rules_jvm_external pin warning
Since we no longer have a single source of truth for maven artifacts, we
need a consistency test. Use some old-school sed to extract the
STARDOC_MAVEN_ARTIFACTS list from MODULE.bazel and deps.bzl, and add a
test to ensure they are consistent. And - for good measure - add a similar test
to ensure the version number in version.bzl and MODULE.bazel is consistent too.

Fixes #174