Improve incompatible flag testing
We always want to use some flags in development; e.g. an empty glob is a
sign of a mistake, so we always want --incompatible_disallow_empty_glob

We will unconditionally set such flags in .bazelrc (not just in CI config).

For other incompatible flags (--incompatible_enable_proto_toolchain_resolution),
we want to test builds both with and without them - and with bzlmod both
enabled and disabled (see discussion in #236 for a case where this
matters).

Such flags we will group under --config=incompatible, and test all 4
compbinations (+- incompatible +- bzlmod) in CI.