Revert most of #237
As the review of #241 showed, --incompatible_enable_proto_toolchain_resolution
is not ready yet; we will re-add support for it in Stardoc when we no
longer need to register our own protoc toolchain in Stardoc's
MODULE.bazel file.

The only parts of #237 we keep are the rules_jvm_external and
rules_proto dep bumps; note that the rules_proto dep bump in turn
requires a new dependency on bazel_features when using a Stardoc distro
release with --noenable_blzmod.